Implement Stefano suggestions
